Permissions Validate

+++ routes/web.php
@@ -15,12 +15,12 @@ Route::get('/dashboard', function () {
 Route::middleware('auth')->group(function () {
 
     Route::prefix('users')->group(function () {
-        Route::get('/', [UsersController::class, 'index'])->name('users.index')->middleware('permission:index users'); // Listar usuários
-        Route::get('/create', [UsersController::class, 'create'])->name('users.create'); // Formulário de criação
-        Route::post('/', [UsersController::class, 'store'])->name('users.store'); // Salvar novo usuário
-        Route::get('/{id}/edit', [UsersController::class, 'edit'])->name('users.edit'); // Formulário de edição
-        Route::put('/{id}', [UsersController::class, 'update'])->name('users.update'); // Atualizar usuário
-        Route::delete('/{id}', [UsersController::class, 'destroy'])->name('users.destroy'); // Excluir usuário
+        Route::get('/', [UsersController::class, 'index'])->name('users.index')->middleware('permission:Lista de usuários'); // Listar usuários
+        Route::get('/create', [UsersController::class, 'create'])->name('users.create')->middleware('permission:Formulário de criação dos usuários'); // Formulário de criação
+        Route::post('/', [UsersController::class, 'store'])->name('users.store')->middleware('permission:Salvar usuários'); // Salvar novo usuário
+        Route::get('/{id}/edit', [UsersController::class, 'edit'])->name('users.edit')->middleware('permission:Formulário de edição dos'); // Formulário de edição
+        Route::put('/{id}', [UsersController::class, 'update'])->name('users.update')->middleware('permission:Atualizar usuários'); // Atualizar usuário
+        Route::delete('/{id}', [UsersController::class, 'destroy'])->name('users.destroy')->middleware('permission:Deletar usuários'); // Excluir usuário
     });
